Refine bulk tag assignment experience
